requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.0.9
djangorestframework==3.11.0
isort==4.3.21
joblib==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.0
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
waitress==1.4.4
whitenoise==5.2.0
wrapt==1.12.1


